+++ packages/expo/ios/Fetch/NativeResponse.swift
@@ -26,7 +26,7 @@ internal final class NativeResponse: SharedObject, ExpoURLSessionTaskDelegate {
   private(set) var responseInit: NativeResponseInit?
   private(set) var redirected = false
   private(set) var error: Error?
-  private(set) var redirectMode: NativeRequestRedirect = .follow
+  var redirectMode: NativeRequestRedirect = .follow
 
   var bodyUsed: Bool {
     return self.sink.bodyUsed
@@ -35,10 +35,6 @@ internal final class NativeResponse: SharedObject, ExpoURLSessionTaskDelegate {
   init(dispatchQueue: DispatchQueue) {
     self.sink = ResponseSink()
     self.dispatchQueue = dispatchQueue
-  }
-
-  func setRedirectMode(_ mode: NativeRequestRedirect) {
-    self.redirectMode = mode
   }
 
   func startStreaming() -> Data? {